First draft clock algorithm for page swapping

+++ src/os345.h
@@ -246,7 +246,7 @@ int lc3Task(int, char**);
 
 unsigned short int *getMemAdr(int va, int rwFlg);
 void outPTE(char* s, int pte);
-long accessPage(int pnum, int frame, int rwnFlg);
+int accessPage(int pnum, int frame, int rwnFlg);
 void initLC3Memory(int startFrame, int endFrame);
 
 #endif // __os345_h__
